+++ example/Nix/nixos-host/configuration.nix
@@ -70,6 +70,7 @@
     };
 
     programs = {
+      dua.enable = true;
       nix.enable = true;
       ranger.enable = true;
       tmux.enable = true;