+++ nixos/modules/programs/default.nix
@@ -1,5 +1,6 @@
 {
   imports = [
+    ./dua.nix
     ./nix.nix
     ./ranger.nix
     ./tmux.nix